+++ src/main/java/de/derivo/neo2rdf/conversion/cli/Neo4jToRDFConversionCLI.java
@@ -11,8 +11,9 @@ import picocli.CommandLine;
                 """,
         subcommands = {Neo4jToTurtleDumpCommand.class, Neo4jToRDFConversionServerCommand.class, CommandLine.HelpCommand.class},
         mixinStandardHelpOptions = true,
+        showDefaultValues = true,
         versionProvider = Neo2RDFVersionProvider.class,
-        usageHelpWidth = 100)
+        usageHelpWidth = 95)
 public class Neo4jToRDFConversionCLI {
 
     public static void main(String[] args) {